Extracted LimitingRowsStrategy enum from SQLStatement (in AbstractSQL) into a standalone class to fix broken modularization and improve class cohesion.
